+++ stroom-search/stroom-search-solr/src/main/java/stroom/search/solr/search/SolrClusterSearchTaskHandler.java
@@ -22,7 +22,7 @@ import org.springframework.stereotype.Component;
 import stroom.annotation.api.AnnotationDataSource;
 import stroom.pipeline.server.errorhandler.MessageUtil;
 import stroom.query.api.v2.ExpressionOperator;
-import stroom.search.coprocessor.CompletionState;
+import stroom.query.common.v2.CompletionState;
 import stroom.search.coprocessor.Coprocessors;
 import stroom.search.coprocessor.CoprocessorsFactory;
 import stroom.search.coprocessor.Error;
@@ -140,7 +140,7 @@ class SolrClusterSearchTaskHandler implements Consumer<Error> {
                 try {
                     taskContext.info("Sending final results");
                     while (!task.isTerminated() && !sendingDataCompletionState.isComplete()) {
-                        sendingDataCompletionState.await(1, TimeUnit.SECONDS);
+                        sendingDataCompletionState.awaitCompletion(1, TimeUnit.SECONDS);
                     }
                 } catch (InterruptedException e) {
                     //Don't want to reset interrupt status as this thread will go back into